fix(fish): use `set -p PATH` instead of `fish_add_path`
The reasoning for this is that `fish_add_path` writes to a universal variable that is persisted independently of fish configuration files, and this can result in stale `fish_user_paths` entries when calls to `fish_add_path` are removed from the configuration. Using `set -p PATH` ensures that only the environment is written to.